chore: move compose example activity to separate module and rename compose->compose-sdk
