+++ .gitignore
@@ -4,7 +4,7 @@
 /sdk/gradle.properties
 /.idea
 .DS_Store
-/build
+build
 /captures
 .externalNativeBuild
 .cxx
