+++ compose-app/build.gradle
@@ -1,18 +1,15 @@
 plugins {
     id 'com.android.application'
+    id 'org.jetbrains.kotlin.android'
 }
 
 def intProp(name, fallback) {
     return project.hasProperty(name) ? Integer.parseInt(project.getProperty(name)) : fallback
 }
 
-def prop(name, fallback) {
-    return project.hasProperty(name) ? project.getProperty(name) : fallback
-}
-
 android {
     compileSdk intProp("exampleCompileSdkVersion", 34)
-    namespace 'com.hcaptcha.example'
+    namespace 'com.hcaptcha.example.compose'
 
     defaultConfig {
         minSdkVersion 23
@@ -21,13 +18,11 @@ android {
         versionName "0.0.1"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        proguardFiles getDefaultProguardFile('proguard-android-optimize.txt')
     }
 
-    compileOptions {
-        // Sets Java compatibility to Java 8
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+    kotlinOptions {
+        jvmTarget = JavaVersion.VERSION_1_8
     }
 
     buildTypes {
@@ -40,14 +35,25 @@ android {
     lint {
         disable 'UsingOnClickInXml'
     }
+
+    buildFeatures { // Enables Jetpack Compose for this module
+        compose = true
+    }
+
+    composeOptions {
+        kotlinCompilerExtensionVersion = "$compose_version"
+    }
 }
 
 dependencies {
     implementation project(path: ':sdk')
+    implementation project(path: ':compose-sdk')
 
-    //noinspection GradleDependency
-    implementation "androidx.appcompat:appcompat:${prop('exampleAppcompatVersion', '1.3.1')}"
-    implementation "com.google.android.flexbox:flexbox:3.0.0"
+    implementation "androidx.compose.ui:ui:$compose_version"
+    implementation 'androidx.compose.material3:material3:1.2.1'
+    implementation 'androidx.activity:activity-ktx:1.8.2'
+    implementation 'androidx.activity:activity-compose:1.8.2'
+    implementation "androidx.compose.foundation:foundation-layout-android:$compose_version"
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
