+++ compose-app/src/main/res/drawable-v24/ic_launcher_foreground.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="108dp"
+    android:height="108dp"
+    android:viewportWidth="480"
+    android:viewportHeight="480">
+    <inset
+        android:insetBottom="16dp"
+        android:insetRight="16dp"
+        android:insetLeft="16dp"
+        android:insetTop="16dp">
+        <path
+            android:pathData="M158.75,222.15l16.8,-37.6c6.12,-9.64 5.31,-21.45 -1.39,-28.15a17.19,17.19 0,0 0,-2.95 -2.36,18.56 18.56,0 0,0 -15.51,-1.81 25.33,25.33 0,0 0,-14 10.83s-23,53.66 -31.56,77.78S105,309.19 138,342.26c35,35 85.68,43 118,18.74a20.27,20.27 0,0 0,3.79 -2.47l99.6,-83.17c4.84,-4 12,-12.24 5.57,-21.64 -6.27,-9.17 -18.15,-2.93 -23,0.17l-57.32,41.68a2.58,2.58 0,0 1,-3.7 -0.43c-1.46,-1.79 -1.72,-6.53 0.57,-8.41l87.87,-74.57c7.59,-6.84 8.65,-16.78 2.5,-23.59 -6,-6.67 -15.53,-6.47 -23.19,0.43l-79.11,61.85a3.51,3.51 0,0 1,-5.06 -0.66c-1.56,-1.75 -2.16,-4.74 -0.4,-6.49l89.6,-86.95a17.52,17.52 0,0 0,0.88 -24.74,16.73 16.73,0 0,0 -12.17,-5.21 17.84,17.84 0,0 0,-12.58 5l-91.55,86c-2.19,2.19 -6.48,0 -7,-2.56a2.79,2.79 0,0 1,0.78 -2.56l70.08,-79.79a17.27,17.27 0,1 0,-24.37 -24.45L171.52,225.93c-3.81,3.81 -9.42,4 -12.09,1.79A4,4 0,0 1,158.75 222.15Z"
+            android:fillColor="#fff"/>
+    </inset>
+</vector>
