+++ compose-sdk/src/androidTest/AndroidManifest.xml
