+++ compose-sdk/src/main/AndroidManifest.xml
