+++ settings.gradle
@@ -3,4 +3,5 @@ include ':sdk'
 include ':test'
 include ':benchmark'
 include ':example-app'
-include ':compose'
+include ':compose-sdk'
+include ':compose-app'
